apps/api/pyproject.toml
[project]
name = "eogum-api"
version = "0.1.0"
description = "어검 (eogum) - Auto Video Edit Online Service API"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "supabase>=2.0.0",
    "PyJWT>=2.8.0",
    "cryptography>=41.0.0",
    "boto3>=1.34.0",
    "httpx>=0.27.0",
    "resend>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.8.0",
]

[project.scripts]
eogum-api = "eogum.main:run"

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 120
